+++ src/main/java/com/marginallyclever/communications/application/ConversationHistoryList.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.robotoverlord.components.robot.robotarm.robotarminterface.marlininterface;
+package com.marginallyclever.communications.application;
 
 import com.marginallyclever.convenience.log.Log;
